Do not remove self-joined users when leaving if they have more sessions
Self-joined users are just temporary participants and they need to be
removed when they leave a conversation. However, this needs to be done
once they have no more sessions in that conversation; otherwise the
remaining sessions would be kicked out from the conversation as soon as
one leaves it.

+++ lib/Service/ParticipantService.php
@@ -713,7 +713,8 @@ class ParticipantService {
 			$this->sessionMapper->deleteByAttendeeId($participant->getAttendee()->getId());
 		}
 
-		if ($participant->getAttendee()->getParticipantType() === Participant::USER_SELF_JOINED) {
+		if ($participant->getAttendee()->getParticipantType() === Participant::USER_SELF_JOINED
+			&& empty($this->sessionMapper->findByAttendeeId($participant->getAttendee()->getId()))) {
 			$this->attendeeMapper->delete($participant->getAttendee());
 
 			$attendeeEvent = new AttendeesRemovedEvent($room, [$participant->getAttendee()]);
